requirements.txt
langchain_community==0.2.5
langchain==0.2.5
langchain_core==0.2.9
pydub==0.25.1
openai==0.28
pyaudio==0.2.14
elevenlabs
python-dotenv
chromadb
pypdf

